Make new books fetch global image. Fix crash.
When a new book is added, a global image (for lists) will be fetched
from the search controller. (The user-provided image will still be
attached and uploaded as per usual, just in the proper location).

When using the camera to scan ISBN, the image (NEW!), author, title, and ISBN
fields will be overwritten with data fetched from the search controller.

Added a ProgressDialog to entertain the user during awkward
pauses after pressing some buttons. (The buttons being Add Book and
opening the ISBN scanner).

Prevented a crash by not letting title, author, or ISBN be empty.
Restricted ISBN to having only 10 or 13 digits.